+++ source/rendering/color2.c
@@ -11,7 +11,7 @@ inline t_color zero_color()
 	return zero;
 }
 
-inline t_color scale_color(t_color v, float scale)
+inline t_color scale_color(t_color v, double scale)
 {
     t_color scaled;
 
@@ -39,7 +39,7 @@ inline t_color sum_colors(t_color amb, t_color dif, t_color   spc)
     return res;
 }
 
-inline float get_brightness(t_color c)
+inline double get_brightness(t_color c)
 {
     return sqrtf( powf(0.299*c.r, 2.f) + powf(0.587*c.g, 2.f) + powf(0.114*c.b, 2.f) );
 }
